# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Claudia Alzate Realtor® API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AuthResult {
  username: string;
}

export interface CredentialsUpdate {
  username?: string;
  password?: string;
  currentPassword?: string;
}

export interface Profile {
  name: string;
  subtitle: string;
  tagline: string;
  /** @nullable */
  logoUrl: string | null;
  /** @nullable */
  backgroundUrl: string | null;
  primaryColor: string;
  goldColor: string;
  fontTitle: string;
  fontBody: string;
}

export interface ProfileUpdate {
  name?: string;
  subtitle?: string;
  tagline?: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  backgroundUrl?: string | null;
  primaryColor?: string;
  goldColor?: string;
  fontTitle?: string;
  fontBody?: string;
}

export interface Link {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  url: string;
  icon: string;
  active: boolean;
  order: number;
}

export interface LinkInput {
  title: string;
  /** @nullable */
  description?: string | null;
  url: string;
  icon: string;
  active?: boolean;
}

export interface LinkUpdate {
  title?: string;
  /** @nullable */
  description?: string | null;
  url?: string;
  icon?: string;
  active?: boolean;
  order?: number;
}

export interface LinksReorder {
  ids: number[];
}