# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Claudia Alzate Realtor® API
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  name: string;
  subtitle: string;
  tagline: string;
  /** @nullable */
  logoUrl: string | null;
  /** @nullable */
  backgroundUrl: string | null;
  primaryColor: string;
  goldColor: string;
  fontTitle: string;
  fontBody: string;
}